[FIXED] iocBadge = null error
Issue: drawShape happens upon creation of shape, and then afterDraw, updateShape can only refer back to shapes that were already drawn.

If the shapes were not previously, there can be a null return when we change IOC to true, i.e. when we drag an IOC node into a combo that was originally IOC false.